# Conflicts:
#	composer.json

+++ application/Espo/Services/Import.php
@@ -3,7 +3,7 @@
  * This file is part of EspoCRM.
  *
  * EspoCRM - Open Source CRM application.
- * Copyright (C) 2014-2017 Yuri Kuznetsov, Taras Machyshyn, Oleksiy Avramenko
+ * Copyright (C) 2014-2018 Yuri Kuznetsov, Taras Machyshyn, Oleksiy Avramenko
  * Website: http://www.espocrm.com
  *
  * EspoCRM is free software: you can redistribute it and/or modify
@@ -38,6 +38,8 @@ use Espo\ORM\Entity;
 
 class Import extends \Espo\Services\Record
 {
+    const REVERT_PERMANENTLY_REMOVE_PERIOD_DAYS = 2;
+
     protected function init()
     {
         parent::init();
@@ -223,6 +225,17 @@ class Import extends \Espo\Services\Record
 
         $sql = "SELECT * FROM import_entity WHERE import_id = ".$pdo->quote($import->id) . " AND is_imported = 1";
 
+        $removeFromDb = false;
+        $createdAt = $import->get('createdAt');
+        if ($createdAt) {
+            $dtNow = new \DateTime();
+            $createdAtDt = new \DateTime($createdAt);
+            $dayDiff = ($dtNow->getTimestamp() - $createdAtDt->getTimestamp()) / 60 / 60 / 24;
+            if ($dayDiff < self::REVERT_PERMANENTLY_REMOVE_PERIOD_DAYS) {
+                $removeFromDb = true;
+            }
+        }
+
         $sth = $pdo->prepare($sql);
         $sth->execute();
         while ($row = $sth->fetch(\PDO::FETCH_ASSOC)) {
@@ -234,12 +247,20 @@ class Import extends \Espo\Services\Record
 
             $entity = $this->getEntityManager()->getEntity($entityType, $entityId);
             if ($entity) {
-                $this->getEntityManager()->removeEntity($entity);
+                $this->getEntityManager()->removeEntity($entity, [
+                    'noStream' => true,
+                    'noNotifications' => true,
+                    'import' => true
+                ]);
             }
-            $this->getEntityManager()->getRepository($entityType)->deleteFromDb($entityId);
+            if ($removeFromDb) {
+                $this->getEntityManager()->getRepository($entityType)->deleteFromDb($entityId);
+            }
         }
 
         $this->getEntityManager()->removeEntity($import);
+
+        $this->processActionHistoryRecord('delete', $import);
 
         return true;
     }
@@ -269,7 +290,7 @@ class Import extends \Espo\Services\Record
             if ($entity) {
                 $this->getEntityManager()->removeEntity($entity);
             }
-            $this->getEntityManager()->getRepository($scope)->deleteFromDb($entityId);
+            $this->getEntityManager()->getRepository($entity->getEntityType())->deleteFromDb($entityId);
         }
 
         return true;
@@ -277,14 +298,14 @@ class Import extends \Espo\Services\Record
 
     public function runIdleImport($data)
     {
-        $entityType = $data['entityType'];
+        $entityType = $data->entityType;
 
-        $params = json_decode(json_encode($data['params']), true);
+        $params = json_decode(json_encode($data->params), true);
 
-        $importFieldList = $data['importFieldList'];
-        $attachmentId = $data['attachmentId'];
+        $importFieldList = $data->importFieldList;
+        $attachmentId = $data->attachmentId;
 
-        $importId = $data['importId'];
+        $importId = $data->importId;
 
         $this->import($entityType, $importFieldList, $attachmentId, $params, $importId);
     }
@@ -326,13 +347,15 @@ class Import extends \Espo\Services\Record
 
         $this->getEntityManager()->saveEntity($import);
 
+        $this->processActionHistoryRecord('create', $import);
+
         if (!empty($params['idleMode'])) {
             $params['idleMode'] = false;
 
             $job = $this->getEntityManager()->getEntity('Job');
             $job->set(array(
                 'serviceName' => 'Import',
-                'method' => 'runIdleImport',
+                'methodName' => 'runIdleImport',
                 'data' => array(
                     'entityType' => $scope,
                     'params' => $params,
@@ -698,11 +721,12 @@ class Import extends \Espo\Services\Record
                 $a[0] = preg_replace('/[^A-Za-z0-9\-]/', '', $a[0]);
 
                 if (count($a) > 1) {
-                    return $a[0] . '.' . $a[1];
+                    return floatval($a[0] . '.' . $a[1]);
                 } else {
-                    return $a[0];
+                    return floatval($a[0]);
                 }
-                break;
+            case Entity::INT:
+                return intval($value);
             case Entity::JSON_OBJECT:
                 $value = \Espo\Core\Utils\Json::decode($value);
                 return $value;
